Initial commit for FreeWave radio module.
Baseline working version with all expected functionality.  Note that the module depends on a constant defined in the RH_Serial dependency for the maximum message length.  The module is based off of the LoRa module; since it also uses RadioHead, implementation is very similar.  Needs to be updated to version 3.0.